@@ -6,9 +6,9 @@
         "@craco/craco": "^7.0.0-alpha.0",
         "bootstrap": "^5.2.3",
         "openai": "^3.2.1",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-bootstrap": "^2.7.4",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-helmet": "^6.1.0",
         "react-router-dom": "^5.2.0"
     },
@@ -34,6 +34,7 @@
         ]
     },
     "devDependencies": {
+        "@testing-library/react": "^14.0.0",
         "react-scripts": "^5.0.1"
     }
 }
