@@ -1,7 +1,6 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-
 import "./style.css";
 import Home from "./components/home/Home";
 
@@ -15,4 +14,9 @@ const App = () => {
     );
 };
 
-ReactDOM.render(<App />, document.getElementById("app"));
+const root = ReactDOM.createRoot(document.getElementById("app"));
+root.render(
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>
+);
